+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [unreleased]
+- [BREAKING] Base path changed from `Rc<Vec<String>>` to `Rc<[String]>`. It means `also Orders::clone_base_path` returns a slice.
 
 ## v0.8.0
 - [BREAKING] Rename `linear_gradient!` to `linearGradient!` for consistency with the other svg macros (same with `radial_gradient!` and `mesh_gradient!`) (#377).